### 2018-03-27 81f0f32
DOC: Fixing issues with paper.md/.bib
