+++ Paper/paper.bib
@@ -26,10 +26,11 @@
             {Torres, Joel},
             {Dell’Acqua, Flavio},
             {Williams, Steve C.R.},
-            {Cash, Diana}}
+            {Cash, Diana}},
   title = {Whole-brain ex-vivo quantitative MRI of the cuprizone mouse model},
   doi = {10.7717/peerj.2632},
-  howpublished = {\url{https://doi.org/10.7717/peerj.2632}}
+  howpublished = {\url{https://doi.org/10.7717/peerj.2632}},
+  year = {2016}
 }
 @paper{Hawkins:2017,
   author = {{Hawkins, Peter C.T.},
@@ -44,10 +45,11 @@
             {Walsh, Eamonn},
             {Mazibuko, Ndabezinhle},
             {Zelaya, Fernando O.},
-            {Mehta, Mitul A.}}
-  title = {An investigation of regional cerebral blood flow and tissue structure changes after acute administration of antipsychotics in healthy male volunteers}
+            {Mehta, Mitul A.}},
+  title = {An investigation of regional cerebral blood flow and tissue structure changes after acute administration of antipsychotics in healthy male volunteers},
   doi = {10.1002/hbm.23844},
-  howpublished = {\url{http://doi.wiley.com/10.1002/hbm.23844}}
+  howpublished = {\url{http://doi.wiley.com/10.1002/hbm.23844}},
+  year = {2017}
 }
 @paper{Richetto:2017,
   author = {{Richetto, Juliet},
@@ -59,17 +61,19 @@
             {Luoni, Alessia},
             {Meyer, Urs},
             {Vernon, Anthony},
-            {Riva, Marco A}}
+            {Riva, Marco A}},
   title = {Genome-Wide Transcriptional Profiling and Structural Magnetic Resonance Imaging in the Maternal Immune Activation Model of Neurodevelopmental Disorders},
   doi = {10.1093/cercor/bhw320},
-  howpublish = {\url{http://dx.doi.org/10.1093/cercor/bhw320}}
+  howpublish = {\url{http://dx.doi.org/10.1093/cercor/bhw320}},
+  year = {2017}
 }
-@paper{Richetto:2017,
+@paper{Vanes:2018,
   author = {{Vanes, Lucy D},
             {Mouchlianitis, Elias},
             {Wood, Tobias C},
-            {Shergill, Sukhi S.}}
+            {Shergill, Sukhi S.}},
   title = {White matter changes in treatment refractory schizophrenia: Does cognitive control and myelination matter?},
   doi = {10.1016/j.nicl.2018.01.010},
-  howpublish = {\url{https://doi.org/10.1016/j.nicl.2018.01.010}}
+  howpublish = {\url{https://doi.org/10.1016/j.nicl.2018.01.010}},
+  year = {2018}
 }
