### 2018-03-27 #652 @spinicist
Hi @arfon and @cmadan - looks like there are some problems with the `paper.md` file. I'll try to diagnose and fix as best I can. Sorry for the inconvenience
