### 2018-03-27 #652 @spinicist
Hello again. I think I have fixed all the issues. One thing that looks odd is that in the affiliation line there is a "1" both before and after the affiliation. I'm not sure what is causing this?

Looking down the list of reviewers, @oesteban is an obvious candidate.

(I'm also happy to volunteer as a reviewer, should I add myself to the Google doc?)
