### 2018-03-27 #652 @arfon
@spinicist - this should fix the affiliation issue: https://github.com/spinicist/QUIT/pull/3

> (I'm also happy to volunteer as a reviewer, should I add myself to the Google doc?)

Great! Please fill out the form here: http://joss.theoj.org/reviewer-signup.html
