### 2018-03-27 8685603
Fixing affiliations (#3)
