### 2018-03-28 #652 @cMadan
@spinicist, this looks like a great project and I'm glad you decided to submit it to JOSS. I also appreciate you adding yourself to the reviewer list :).

@oesteban, thanks for volunteering to review--and so promptly!

Let's start the review! :)
