### 2018-03-28 #656 @whedon
[REVIEW]: Quantitative Imaging Tools
**Submitting author:** @spinicist (<a href="http://orcid.org/0000-0001-7640-5520">Tobias Wood</a>)
**Repository:** <a href="https://github.com/spinicist/QUIT" target ="_blank">https://github.com/spinicist/QUIT</a>
**Version:** v2.0.0
