### 2018-04-09 #656 @spinicist
Hello, I've been away but am now back. Is there anything I need to do for this except wait? The article proof looks okay to me. Thanks.
