### 2018-04-09 #656 @oesteban
I made a first pass over the conformity checks. I hope I'll go through installation and the rest along this week.

I guess all is on my plate right now
