### 2018-04-10 2fd57fc
ENH: Bumped ceres version, improved easy_build.sh
