+++ easy_build.sh
@@ -13,6 +13,28 @@
 #
 
 WD=$PWD
+
+# Command line options
+QUIT_INSTALL_DIR="" # User can specify install directory (mainly for Travis)
+NUM_THREADS="2"     # Specify number of threads during make/ninja
+NATIVE=""           # Specifieds march=native
+while getopts "p:j:n" opt; do
+    case $opt in
+        p) QUIT_INSTALL_DIR="-DCMAKE_INSTALL_PREFIX=$OPTARG";;
+        j) NUM_THREADS="$OPTARG";;
+        n) NATIVE="-DCMAKE_CXX_FLAGS=-march=native";;
+    esac
+done
+
+# Check for presence of ninja
+if [ -x "$(which ninja)" ]; then
+    GENERATOR="-GNinja"
+    BUILDCMD="ninja -j $NUM_THREADS"
+else
+    GENERATOR=""
+    BUILDCMD="make -j $NUM_THREADS"
+fi
+
 # Initialise submodules
 git submodule update --init
 EXTERNAL="$WD/External"
@@ -23,23 +45,14 @@ EIGEN_DIR="$EXTERNAL/eigen"
 ARGS_DIR="$EXTERNAL/args"
 CEREAL_DIR="$EXTERNAL/cereal/include"
 
-# Check for presence of ninja
-if [ -x "$(which ninja)" ]; then
-    GENERATOR="-GNinja"
-    BUILDCMD="ninja -j 2"
-else
-    GENERATOR=""
-    BUILDCMD="make -j 2"
-fi
-
 # Ceres
 CERES_DIR="$EXTERNAL/ceres-solver"
-CERES_BUILD_DIR="$CERES_DIR/build"
-CERES_OPTS="$GENERATOR -DCMAKE_BUILD_TYPE=Release\
+CERES_BUILD_DIR="$CERES_DIR/cmake_build"
+CERES_OPTS="$GENERATOR -DCMAKE_BUILD_TYPE=Release $NATIVE\
             -DBUILD_DOCUMENTATION=OFF -DBUILD_EXAMPLES=OFF\
             -DBUILD_SHARED_LIBS=OFF -DBUILD_TESTING=OFF\
             -DCUSTOM_BLAS=OFF -DCXSPARSE=OFF -DCXX11=ON\
-            -DEIGEN_INCLUDE_DIR=${EIGEN_DIR} -DEIGENSPARSE=OFF\
+            -DEIGEN_INCLUDE_DIR=$EIGEN_DIR -DEIGENSPARSE=OFF\
             -DEXPORT_BUILD_DIR=ON\
             -DGFLAGS=OFF -DLAPACK=OFF -DMINIGLOG=ON\
             -DOPENMP=OFF -DSUITESPARSE=OFF"
@@ -50,10 +63,12 @@ $BUILDCMD
 
 # ITK
 ITK_DIR="$EXTERNAL/ITK"
-ITK_BUILD_DIR="$ITK_DIR/build"
-ITK_OPTS="$GENERATOR -DCMAKE_BUILD_TYPE=Release -DBUILD_EXAMPLES=OFF\
-           -DBUILD_TESTING=OFF -DITK_BUILD_DEFAULT_MODULES=OFF -DITKGroup_Core=OFF\
-           -DModule_ITKCommon=ON -DModule_ITKIONIFTI=ON -DModule_ITKIONRRD=ON\
+ITK_BUILD_DIR="$ITK_DIR/cmake_build"
+ITK_OPTS="$GENERATOR -DCMAKE_BUILD_TYPE=Release $NATIVE
+           -DBUILD_EXAMPLES=OFF -DBUILD_TESTING=OFF\
[…]
