### 2018-04-10 6a4c021
BUG: Fixes for finding Eigen
