+++ easy_build.sh
@@ -52,7 +52,7 @@ CERES_OPTS="$GENERATOR -DCMAKE_BUILD_TYPE=Release $NATIVE\
             -DBUILD_DOCUMENTATION=OFF -DBUILD_EXAMPLES=OFF\
             -DBUILD_SHARED_LIBS=OFF -DBUILD_TESTING=OFF\
             -DCUSTOM_BLAS=OFF -DCXSPARSE=OFF -DCXX11=ON\
-            -DEIGEN_INCLUDE_DIR=$EIGEN_DIR -DEIGENSPARSE=OFF\
+            -DEIGEN_INCLUDE_DIR_HINTS=$EIGEN_DIR -DEIGENSPARSE=OFF\
             -DEXPORT_BUILD_DIR=ON\
             -DGFLAGS=OFF -DLAPACK=OFF -DMINIGLOG=ON\
             -DOPENMP=OFF -DSUITESPARSE=OFF"
@@ -101,7 +101,7 @@ cd $WD
 
 QUIT_BLD_DIR="cmake_build"
 QUIT_OPTS="$GENERATOR -DCMAKE_BUILD_TYPE=Release $NATIVE\
-    -DEIGEN3_INCLUDE_DIR=$EIGEN_DIR -DCeres_DIR=$CERES_BUILD_DIR\
+    -DEIGEN3_INCLUDE_DIRS=$EIGEN_DIR -DCeres_DIR=$CERES_BUILD_DIR\
     -DITK_DIR=$ITK_BUILD_DIR -DArgs_DIR=$ARGS_DIR -DCereal_DIR=$CEREAL_DIR
     $QUIT_INSTALL_DIR"
 mkdir -p $QUIT_BLD_DIR
