### 2018-04-12 4cc19ff
DOC: Fixed some typos
