### 2018-04-17 c0b103d
ENH: Improvements to ellipse processing
Swapped from DynamicAutoDiffCostFunction to AutoDiffCostFunction.
General tweaks to formulas/starting positions.
Added decent debugging messages.
