+++ Source/SSFP/DirectAlgo.cpp
@@ -24,23 +24,27 @@ public:
     const bool debug;
 
     template<typename T>
-    bool operator() (T const* const* p, T* resids) const {
+    bool operator() (const T *const p, T* resids) const {
         typedef Eigen::Array<T, Eigen::Dynamic, 1> ArrayXT;
 
-        const T &G = p[0][0];
-        const T &a = p[0][1];
-        const T &b = p[0][2];
-        const T &th0 = p[0][3];
-        const T &psi0 = p[0][4];
+        const T &G = p[0];
+        const T &a = p[1];
+        const T &b = p[2];
+        const T &th0 = p[3];
+        const T &psi0 = p[4];
 
         if (b < 2.*a/(1. + a*a)) {
             ArrayXT m = EllipseToSignal(G, a, b, th0, psi0, TR, phi);
             Eigen::Map<ArrayXT> r(resids, data.size()*2);
             r.head(data.size()) = m.head(data.size()) - data.real();
             r.tail(data.size()) = m.tail(data.size()) - data.imag();
+            Eigen::IOFormat numpy(4, 0, ", ", ",", "[", "]", "np.array(", ")");
             if (debug) {
-                std::cout << "*** ELLIPSE COST *** "
-                    << "G " << G << " a " << a << " b " << b << " th0 " << th0 << " psi0 " << psi0 << std::endl;
+                Eigen::ArrayXd md(2*data.size()); for (int i = 0; i < 2*data.size(); i++) { md[i] = m[i].a; }
+                std::cout 
+                    << "{ 'P': { 'G': " << G.a << ", 'a': " << a.a << " , 'b': " << b.a << " , 'th0': " << th0.a << " , 'psi0': " << psi0.a << "},"
+                    << " 'M': { 'real': " << md.head(data.size()).transpose().format(numpy) << ", 'imag': " << md.tail(data.size()).transpose().format(numpy) << "},"
+                    << " 'D': { 'real': " << data.real().transpose().format(numpy) << ", 'imag': " << data.imag().transpose().format(numpy) << "}}," << std::endl;
             }
             return true;
         } else {
@@ -53,23 +57,50 @@ public:
     }
 };
 
-Eigen::ArrayXd DirectAlgo::apply_internal(const Eigen::ArrayXcf &indata, const double flip, const double TR, const Eigen::ArrayXd &phi, const bool debug, float &residual) const {
+    template<>
+    bool DirectCost::operator() <double> (const double *const p, double *resids) const {
+        const double &G = p[0];
+        const double &a = p[1];
+        const double &b = p[2];
+        const double &th0 = p[3];
+        const double &psi0 = p[4];
+        
+        if (b < 2.*a/(1. + a*a)) {
+            Eigen::ArrayXd m = EllipseToSignal(G, a, b, th0, psi0, TR, phi);
+            Eigen::Map<Eigen::ArrayXd> r(resids, data.size()*2);
+            r.head(data.size()) = m.head(data.size()) - data.real();
+            r.tail(data.size()) = m.tail(data.size()) - data.imag();
+            if (debug) {
+
+            }
+            return true;
+        } else {
+            if (debug) {
+                std::cout << "*** ELLIPSE COST NOT EVALUATED - ELLIPSE WAS HORIZONTAL *** " 
+                    << "G " << G << " a " << a << " b " << b << " th0 " << th0 << " psi0 " << psi0 << std::endl;
+            }
+            return false;
+        }
+    }
+
+
+Eigen::ArrayXd DirectAlgo::apply_internal(const Eigen::ArrayXcf &indata,
[…]
