### 2018-04-20 0343a6b
BUG: Lower/upper bounds were not JSON
