### 2018-04-24 84228ab
ENH: Polynomial fitting improvements
Added scaling to improve numerical stability for high-order polys
JSONified input/output
