+++ Source/Core/Fit.h
@@ -18,8 +18,8 @@
 
 namespace QI {
 
-Eigen::VectorXd LeastSquares(const Eigen::MatrixXd &X, const Eigen::VectorXd &y);
-Eigen::VectorXd RobustLeastSquares(const Eigen::MatrixXd &X, const Eigen::VectorXd &y);
+Eigen::VectorXd LeastSquares(const Eigen::MatrixXd &X, const Eigen::VectorXd &y, double *resid = nullptr);
+Eigen::VectorXd RobustLeastSquares(const Eigen::MatrixXd &X, const Eigen::VectorXd &y, double *resid = nullptr);
 
 } // End namespace QI
 
