### 2018-04-25 cccb99c
BUG: Found a few typos in example scripts
