+++ mkdocs.yml
@@ -2,6 +2,7 @@ site_name: QUIT
 site_description: Documentation for the QUantitative Imaging Tools
 site_url: https://spinicist.github.io/QUIT
 repo_url: https://github.com/spinicist/QUIT
+edit_uri: ""
 
 docs_dir: Docs
 theme: mkdocs
