### 2018-04-25 3673e42
DOC: Fix link to releases page
