+++ Docs/Relaxometry.md
@@ -4,17 +4,17 @@ Relaxometry is the measurement of the longitudinal and transverse relaxation tim
 
 The following programs are available:
 
-* [qidespot1](##qidespot1)
-* [qidespot1hifi](##qidespot1hifi)
-* [qidespot2](##qidespot2)
-* [qidespot2fm](##qidespot2fm)
-* [qimcdespot](##qimcdespot)
-* [qimultiecho](##qimultiecho)
-* [qimp2rage](##qimp2rage)
-* [qiafi](##qiafi)
-* [qidream](##qidream)
+* [qidespot1](#qidespot1)
+* [qidespot1hifi](#qidespot1hifi)
+* [qidespot2](#qidespot2)
+* [qidespot2fm](#qidespot2fm)
+* [qimcdespot](#qimcdespot)
+* [qimultiecho](#qimultiecho)
+* [qimp2rage](#qimp2rage)
+* [qiafi](#qiafi)
+* [qidream](#qidream)
 
-## qidespot1
+##qidespot1
 
 This program implements the classic Driven Equilibrium Single-Pulse Observation of T1 (DESPOT1) algorithm, also known as the Variable Flip-Angle (VFA) method in the literature. This is a fast way to measure longitudinal relaxation using a spoiled steady-state sequence, which is know by a different name by every scanner manufacturer just to be helpful. On GE, it's SPoiled Gradient Recalled echo (SPGR), on Siemens it's Fast Low-Angle SHot (FLASH) and on Phillips the sequence is Fast Field Echo (FFE).
 
@@ -62,7 +62,7 @@ qidespot1 input_file.nii.gz --mask=mask_file.nii.gz --B1=b1_file.nii.gz < input.
 [2]: http://doi.wiley.com/10.1002/mrm.21669
 [3]: http://doi.wiley.com/10.1002/mrm.25592
 
-## qidespot1hifi
+##qidespot1hifi
 
 This is an extension of DESPOT1 to fit a map simultaneously using an MP-RAGE / IR-SPGR type sequence. Although DESPOT1-HIFI can produce a rough estimate of B1, it often fails to produce reasonable values in the ventricles, and the fact that the MP-RAGE image is often acquired at lower resolution than the SPGR/FLASH data can also cause problems. Hence you should either smooth the B1 map produced as output, or fit it with a [polynomial](Utilities.md), then recalculate T1 using the [DESPOT1](##qidespot1) program. Note that if your MP-RAGE image is not acquired at the same resolution as your SPGR data, it must be resampled to the same spacing before processing (and it should also be registered to your SPGR data).
 
@@ -106,7 +106,7 @@ For the MPRAGE sequence, the TR is the spacing between readouts/echoes, not the 
 
 [1]: http://doi.wiley.com/10.1002/jmri.21130
 
-## qidespot2
+##qidespot2
 
 DESPOT2 uses SSFP data and a separate T<sub>1</sub> map to calculate T<sub>2</sub>, using the same maths as DESPOT1. It does not account for the banding artefacts present in SSFP data at field-strengths of 3T and above. See [qidespot2fm](#qidespot2fm) for a method that does account for them, or if you have at least 4 phase-increments and complex data then see [qi_ssfp_bands](SSFP.md#qi_ssfp_bands) for a way to remove them before using this program.
 
@@ -157,7 +157,7 @@ Both `PhaseInc` and `FA` are measured in degrees. If the ellipse option is speci
 
 [1]: http://doi.wiley.com/10.1002/mrm.10407
 
-## qidespot2fm
+##qidespot2fm
 
[…]
