### 2018-04-26 8da6849
BUG: More typos in example script
