### 2018-04-26 3048536
BUG: qiaffine was mangling non-isotropic voxel spacings
