+++ Source/Utils/qiaffine.cpp
@@ -15,6 +15,7 @@
 #include "itkVersor.h"
 #include "itkVersorRigid3DTransform.h"
 #include "itkCenteredAffineTransform.h"
+#include "itkEuler3DTransform.h"
 #include "itkTransformFileWriter.h"
 #include "itkPermuteAxesImageFilter.h"
 #include "itkFlipImageFilter.h"
@@ -23,8 +24,6 @@
 #include "Util.h"
 #include "ImageIO.h"
 #include "Args.h"
-
-using namespace std;
 
 /*
  * Declare arguments here so they are available to pipeline
@@ -41,15 +40,11 @@ args::HelpFlag help(parser, "HELP", "Show this help menu", {'h', "help"});
 args::Flag     verbose(parser, "VERBOSE", "Print more information", {'v', "verbose"});
 args::ValueFlag<std::string> center(parser, "CENTER", "Set the origin to geometric center (geo) or (cog)", {'c', "center"});
 args::ValueFlag<std::string> tfm_path(parser, "TFM", "Write out the transformation to a file", {'t', "tfm"});
+args::ValueFlag<std::string> permute(parser, "PERMUTE", "Permute axes in data-space, e.g. 2,0,1. Negative values mean flip as well", {"permute"});
+args::ValueFlag<std::string> flip(parser, "FLIP", "Flip axes in data-space, e.g. 0,1,0. Occurs AFTER any permutation.", {"flip"});
 args::ValueFlag<double> scale(parser, "SCALE", "Scale by a constant", {'s', "scale"});
-args::ValueFlag<double> offX(parser, "OFF_X", "Translate origin in X direction", {"offX"});
-args::ValueFlag<double> offY(parser, "OFF_Y", "Translate origin in Y direction", {"offY"});
-args::ValueFlag<double> offZ(parser, "OFF_Z", "Translate origin in Z direction", {"offZ"});
-args::ValueFlag<double> rotX(parser, "ROT_X", "Rotate about X-axis by angle (degrees)", {"rotX"});
-args::ValueFlag<double> rotY(parser, "ROT_Y", "Rotate about Y-axis by angle (degrees)", {"rotY"});
-args::ValueFlag<double> rotZ(parser, "ROT_Z", "Rotate about Z-axis by angle (degrees)", {"rotZ"});
-args::ValueFlag<std::string> permute(parser, "PERMUTE", "Permute axes, e.g. 2,0,1. Negative values mean flip as well", {"permute"});
-args::ValueFlag<std::string> flip(parser, "FLIP", "Flip an axis, e.g. 0,1,0. Occurs AFTER any permutation.", {"flip"});
+args::ValueFlag<std::string> translate(parser, "TRANSLATE", "Translate image by X,Y,Z (mm)", {"trans"}, "0,0,0");
+args::ValueFlag<std::string> rotate(parser, "ROTATE", "Rotate by Euler angles around X,Y,Z (degrees).", {"rotate"}, "0,0,0");
 
 template<typename TImage>
 int Pipeline() {
@@ -110,8 +105,9 @@ int Pipeline() {
     QI::VolumeF::DirectionType direction;
     QI::VolumeF::SpacingType spacing;
 
-    typedef itk::CenteredAffineTransform<double, 3> TAffine; 
-    TAffine::OutputVectorType origin;
+    using Affine = itk::CenteredAffineTransform<double, 3>; 
+    using Euler  = itk::Euler3DTransform<double>;
+    Affine::OutputVectorType origin;
     QI::VolumeF::SizeType size;
     for (int i = 0; i < 3; i++) {
         for (int j = 0; j < 3; j++) {
@@ -122,30 +118,28 @@ int Pipeline() {
[…]
