### 2018-04-26 6ce7db8
BUG: Lower bound on qidespot2fm f0 not set correctly
