### 2018-04-27 f35227b
BUG: 3048536 introduced a scaling bug in qiaffine
