### 2018-04-28 spinicist/QUIT#6 @oesteban
[JOSS Review] Packaging
@spinicist, have you considered to create a debian package and submit it to neurodebian?

(ref openjournals/joss-reviews#656)
