### 2018-04-28 spinicist/QUIT#7 @oesteban
[JOSS Review] Community guidelines
Please add community guidelines. Now it is really easy with GitHub. Adding neurostars.org as a support channel would also increase the visibility of these tools.

(ref openjournals/joss-reviews#656)
