### 2018-05-08 daf7739
DOC: Added CONTRIBUTING.md
