### 2018-05-08 38ae23e
DOC: Updated changelog, added a getting help section to readme
