### 2018-05-08 3f4039f
ENH: Added Debian CPack generator STYLE: Enabled -Wall and -Wextra
Added SYSTEM flag to include_directories() calls to stop warnings
being generated for Args/Cereal/Eigen/Ceres. Currently don't know
the equivalent for ITK, so have not added -Wpedantic as that
generates a lot of warnings from ITK macros. Enabling warnings
threw up a LOT of signed/unsigned warnings, several unused variable
warnings and one uninitialized variable.
