### 2018-05-09 ce9ac08
STYLE: GCC picked up an old XCode pragma
