+++ CMakeLists.txt
@@ -41,7 +41,7 @@ OPTION(BUILD_SHARED_LIBS "Build shared libraries." OFF)
 
 set( VERSION_MAJOR "2" )
 set( VERSION_MINOR "0" )
-set( VERSION_PATCH "0" )
+set( VERSION_PATCH "1" )
 
 add_subdirectory( Source )
 add_subdirectory( Scripts )
