### 2018-05-09 spinicist/QUIT#7 @spinicist
I have added community guidelines in `CONTRIBUTING.md`, which I think Github should automatically pick up for new Pull Requests.

I have joined Neurostars (I had been meaning to do that anyway) and have mentioned this is in `README.md`. Is there anything else I need to do to make this a 'support channel' (i.e. do I need to advertise on Neurostars somehow?)

There's been a lot of updates to the documentation and some bug fixes so I have released v2.0.1 as well.
