### 2018-05-09 spinicist/QUIT#6 @oesteban
This is excellent news!
