### 2018-05-14 d66515e
BUG: Reverted to CMake 3.0 for building Debian
