### 2018-05-16 4f5fb41
ENH Improves Nipype interfaces
Adds more interfaces
Changes module name from Qi to QI
Adds example notebook with simulations
Adds setup.py
