+++ Python/QUIT/utils.py
@@ -1,0 +1,299 @@
+#! /usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+Implementation of nipype interfaces for QUIT utilities.
+
+Implemented:
+    - qiaffine
+    - qimask
+    - qi_rfprofile
+    
+To be implemented:
+    - qi_coil_combine
+    - qicomplex
+    - qihdr
+    - qikfilter
+    - qipolyfit
+    - qipolyimg
+
+Requires that the QUIT tools are in your your system path
+"""
+
+from __future__ import (print_function, division, unicode_literals,
+                        absolute_import)
+
+from nipype.interfaces.base import CommandLineInputSpec, CommandLine, TraitedSpec, File, traits, isdefined
+# from quit_nipype_utils import parse_param_dict
+from .base import QUITCommand, QUITCommandInputSpec
+
+import json
+import os
+
+############################ qi_rfprofile ############################
+
+
+class QiRFprofileInputSpec(QUITCommandInputSpec):
+
+    # Inputs
+    in_file = File(exists=True, argstr='%s', mandatory=True,
+                   position=0, desc='Input B1+ file')
+    param_file = File(desc='Parameter .json file', position=2, argstr='< %s',
+                      xor=['param_dict'], mandatory=True, exists=True)
+    param_dict = traits.Dict(desc='Parameter dictionary', position=2,
+                             argstr='', mandatory=True, xor=['param_file'])
+
+    # Outputs
+    out_file = File(exists=False, argstr='%s', mandatory=True,
+                    desc='Input B1+ file', position=1)
+
+    # Options
+    mask_file = File(
+        desc='Only process voxels within the mask', argstr='--mask=%s')
+
+    # Commonly used options
+    debug = traits.Bool(desc='Output debugging messages', argstr='-v')
+    threads = traits.Int(
+        desc='Use N threads (default=4, 0=hardware limit)', argstr='--threads=%d')
+    prefix = traits.String(
+        desc='Add a prefix to output filenames', argstr='--out=%s')
+
+
+class QiRFprofileOutputSpec(TraitedSpec):
+    # Specify which outputs there are
+    out_file = File(desc="output relative B1 file")
+
+
+class QiRFprofile(QUITCommand):
+    """
+    help for myInterface
+
+    Example 1
+    -------
+    >>> from QUIT.nipype.utils import QiRFprofile
+    >>> interface = QiRFprofile()
+
+    """
+
+    _cmd = 'qi_rfprofile'
+    input_spec = QiRFprofileInputSpec
+    output_spec = QiRFprofileOutputSpec
+
+    def _format_arg(self, name, spec, value):
+        return self._process_params(name, spec, value)
+
+    def _list_outputs(self):
+        outputs = self.output_spec().get()
+
+        outputs['out_file'] = os.path.abspath(
+            self._add_prefix(self.inputs.out_file))
+
+        return outputs
+
+############################ qiaffine ############################
+
+
+class QiAffineInputSpec(QUITCommandInputSpec):
+
+    # Inputs
+    in_file = File(exists=True, argstr='%s', mandatory=True,
+                   position=0, desc='Source File')
+
+    # Outputs
+    out_file = File(exists=False, argstr='%s', mandatory=True,
+                    position=1, desc='Destination File')
+
+    # Options
+    xfm_file = File(exists=False, argstr='--tfm=%s', mandatory=True,
+                    desc='Write out the transformation to a file')
+    offX = traits.Int(desc='Translate origin in X direction',
+                      argstr='-offX=%d')
+    offY = traits.Int(desc='Translate origin in Y direction',
+                      argstr='-offY=%d')
+    offZ = traits.Int(desc='Translate origin in Z direction',
+                      argstr='-offZ=%d')
[…]
