+++ Python/QUIT/ssfp.py
@@ -1,0 +1,27 @@
+#! /usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+Implementation of nipype interfaces for QUIT SSFP modules.
+
+To be implemented:
+    - qi_ssfp_bands
+    - qi_ssfp_elipse
+    - qi_ssfp_emt
+    - qi_ssfp_planet
+
+Requires that the QUIT tools are in your your system path
+
+"""
+
+############################ qi_ssfp_bands ############################
+# < To be implemented > #
+
+############################ qi_ssfp_elipse ############################
+# < To be implemented > #
+
+############################ qi_ssfp_emt ############################
+# < To be implemented > #
+
+############################ qi_ssfp_planet ############################
+# < To be implemented > #
