+++ Python/QUIT/perfusion.py
@@ -1,0 +1,23 @@
+#! /usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+Implementation of nipype interfaces for QUIT Perfusion modules.
+
+To be implemented:
+    - qi_ase_oef
+    - qi_asl
+    - qi_zshim
+
+Requires that the QUIT tools are in your your system path
+
+"""
+
+############################ qi_ase_oef ############################
+# < To be implemented > #
+
+############################ qi_asl ############################
+# < To be implemented > #
+
+############################ qi_zshim ############################
+# < To be implemented > #
