### 2018-05-16 a0684eb
Nipype interface for qidespot1
