### 2018-05-17 f8ae24e
DOC: Changed all .txt input-file extensions to .json
