### 2018-05-17 517c2b3
BUG: Increased k_f upper bound, improves fitting result
