### 2018-05-17 07b5f6c
BUG: Added extra Update() call to get correct region size
The unpadding step was failing for images that needed padding.
This was because the correct region size info was not available
