### 2018-05-17 961b45f
BUG: qidespot2 was not printing help message with no arguments
