### 2018-05-17 54aa518
BUG: Umesh Rudrapatna reported ARLO formula in the paper is incorrect
