+++ Source/CoreProgs/qidiff.cpp
@@ -31,7 +31,7 @@ int main(int argc, char **argv) {
     args::ValueFlag<std::string> input_path(parser, "INPUT", "Input file for difference", {"input"});
     args::ValueFlag<std::string> baseline_path(parser, "BASELINE", "Baseline file for difference", {"baseline"});
     args::ValueFlag<double>      tolerance(parser, "TOLERANCE", "Tolerance (mean percent difference)", {"tolerance"}, 0);
-    args::ValueFlag<double>      noise(parser, "NOISE", "Added noise level, tolerance is relative to this", {"noise"}, 1);
+    args::ValueFlag<double>      noise(parser, "NOISE", "Added noise level, tolerance is relative to this", {"noise"}, 0);
     args::Flag                   absolute(parser, "ABSOLUTE", "Use absolute difference, not relative (avoids 0/0 problems)", {'a', "abs"});
     QI::ParseArgs(parser, argc, argv, verbose);
     if (verbose) std::cout << "Reading input: " << QI::CheckValue(input_path) << std::endl;
@@ -59,7 +59,7 @@ int main(int argc, char **argv) {
 
     const double mean_sqr_diff = stats->GetMean();
     const double root_mean_sqr_diff = sqrt(mean_sqr_diff);
-    const double rel_diff = root_mean_sqr_diff / noise.Get();
+    const double rel_diff = (noise.Get() > 0) ? root_mean_sqr_diff / noise.Get() : root_mean_sqr_diff;
     const bool passed = rel_diff <= tolerance.Get();
     if (verbose) {
         std::cout << "Mean Square Diff: " << mean_sqr_diff
