### 2018-05-17 1914ad3
WIP: Added qidespot2 to nipype
