+++ Python/QUIT/nipype/relaxometry.py
@@ -38,10 +38,10 @@ class QiDespot1InputSpec(CommandLineInputSpec):
     b1map_file = File(desc='B1 map (ratio) file', argstr='--B1=%s')
     mask_file = File(desc='Only process voxels within the mask', argstr='--mask=%s')
     residuals = traits.Bool(desc='Write out residuals for each data-point', argstr='--resids')
-    algo = traits.String(desc="Choose algorithm (l/w/n)", argstr="--algo=%s")
+    algo = traits.Enum("LLS","WLS","NLS", desc="Choose algorithm", argstr="--algo=%d")
     iterations = traits.Int(desc='Max iterations for WLLS/NLLS (default 15)', argstr='--its=%d')
     clamp_PD = traits.Float(desc='Clamp PD between 0 and value', argstr='-f %f')
-    clamp_T1 = traits.Int(desc='Clamp T1 between 0 and value', argstr='--clampT1=%f')
+    clamp_T1 = traits.Float(desc='Clamp T1 between 0 and value', argstr='--clampT1=%f')
     environ = {'QUIT_EXT':'NIFTI_GZ'}
     
 class QiDespot1OutputSpec(TraitedSpec):
@@ -96,5 +96,92 @@ class QiDespot1(CommandLine):
         
         if self.inputs.residuals:
             outputs['residual_map'] = prefix + 'D1_residual.nii.gz'
+        
+        return outputs
+
+############################ qidespot2 ############################
+
+class QiDespot2InputSpec(CommandLineInputSpec):
+    
+    # Inputs
+    param_file = File(desc='Parameter .json file', position=-1, argstr='< %s', 
+        xor=['param_dict'], mandatory=True, exists=True)
+
+    param_dict = traits.Dict(desc='dictionary trait', position=-1, 
+        argstr='', mandatory=True, xor=['param_file'])
+
+    t1_map = File(exists=True, argstr='%s', mandatory=True,
+        position=-3, desc='Path to T1 map')
+    in_file = File(exists=True, argstr='%s', mandatory=True,
+        position=-2, desc='Path to SSFP data')
+
+    # Options
+    verbose = traits.Bool(desc='Print more information', argstr='-v')
+    threads = traits.Int(desc='Use N threads (default=4, 0=hardware limit)', argstr='--threads=%d')
+    prefix = traits.String(desc='Add a prefix to output filenames', argstr='--out=%s')
+    b1map_file = File(desc='B1 map (ratio) file', argstr='--B1=%s')
+    mask_file = File(desc='Only process voxels within the mask', argstr='--mask=%s')
+    residuals = traits.Bool(desc='Write out residuals for each data-point', argstr='--resids')
+    algo = traits.Enum("LLS","WLS","NLS", desc="Choose algorithm", argstr="--algo=%d")
+    ellipse = traits.Bool(desc="Data is ellipse geometric solution", argstr='-e')
+    iterations = traits.Int(desc='Max iterations for WLLS/NLLS (default 15)', argstr='--its=%d')
+    clamp_PD = traits.Float(desc='Clamp PD between 0 and value', argstr='-f %f')
+    clamp_T2 = traits.Float(desc='Clamp T2 between 0 and value', argstr='--clampT1=%f')
+    environ = {'QUIT_EXT':'NIFTI_GZ'}
+    
+class QiDespot2OutputSpec(TraitedSpec):
+    t2_map = File(desc="Path to T2 map")
+    pd_map = File(desc="Path to PD map")
+    residual_map = File(desc="Path to residual map")
+
+class QiDespot2(CommandLine):
+    """
[…]
