+++ Python/QUIT/nipype/relaxometry.py
@@ -20,17 +20,16 @@ import json
 ############################ qidespot1 ############################
 
 class QiDespot1InputSpec(CommandLineInputSpec):
-    
     # Inputs
-    param_file = File(desc='Parameter .json file', position=-1, argstr='< %s', 
+    spgr_file = File(exists=True, argstr='%s', mandatory=True,
+        position=0, desc='Path to SPGR data')
+
+    param_file = File(desc='Parameter .json file', position=1, argstr='< %s', 
         xor=['param_dict'], mandatory=True, exists=True)
 
-    param_dict = traits.Dict(desc='dictionary trait', position=-1, 
+    param_dict = traits.Dict(desc='dictionary trait', position=1, 
         argstr='', mandatory=True, xor=['param_file'])
-        
-    in_file = File(exists=True, argstr='%s', mandatory=True,
-        position=-2, desc='Path to SPGR data')
-    
+
     # Options
     verbose = traits.Bool(desc='Print more information', argstr='-v')
     threads = traits.Int(desc='Use N threads (default=4, 0=hardware limit)', argstr='--threads=%d')
@@ -38,10 +37,10 @@ class QiDespot1InputSpec(CommandLineInputSpec):
     b1map_file = File(desc='B1 map (ratio) file', argstr='--B1=%s')
     mask_file = File(desc='Only process voxels within the mask', argstr='--mask=%s')
     residuals = traits.Bool(desc='Write out residuals for each data-point', argstr='--resids')
-    algo = traits.String(desc="Choose algorithm (l/w/n)", argstr="--algo=%s")
+    algo = traits.Enum("LLS","WLS","NLS", desc="Choose algorithm", argstr="--algo=%d")
     iterations = traits.Int(desc='Max iterations for WLLS/NLLS (default 15)', argstr='--its=%d')
     clamp_PD = traits.Float(desc='Clamp PD between 0 and value', argstr='-f %f')
-    clamp_T1 = traits.Int(desc='Clamp T1 between 0 and value', argstr='--clampT1=%f')
+    clamp_T1 = traits.Float(desc='Clamp T1 between 0 and value', argstr='--clampT1=%f')
     environ = {'QUIT_EXT':'NIFTI_GZ'}
     
 class QiDespot1OutputSpec(TraitedSpec):
@@ -96,5 +95,317 @@ class QiDespot1(CommandLine):
         
         if self.inputs.residuals:
             outputs['residual_map'] = prefix + 'D1_residual.nii.gz'
+        
+        return outputs
+
+############################ qidespot1hifi ############################
+
+class QiDespot1HifiInputSpec(CommandLineInputSpec):
+    # Inputs
+    spgr_file = File(exists=True, argstr='%s', mandatory=True,
+        position=0, desc='Path to SPGR data')
+
+    mprage_file = File(exists=True, argstr='%s', mandatory=True,
+        position=0, desc='Path to MPRAGE data')
+
+    param_file = File(desc='Parameter .json file', position=2, argstr='< %s', 
+        xor=['param_dict'], mandatory=True, exists=True)
+
+    param_dict = traits.Dict(desc='dictionary trait', position=2, 
+        argstr='', mandatory=True, xor=['param_file'])
+
+    # Options
+    verbose = traits.Bool(desc='Print more information', argstr='-v')
+    threads = traits.Int(desc='Use N threads (default=4, 0=hardware limit)', argstr='--threads=%d')
[…]
