### 2018-05-17 33a6aea
WIP: Added HIFI to nipype
