+++ Python/QUIT/nipype/relaxometry.py
@@ -20,17 +20,16 @@ import json
 ############################ qidespot1 ############################
 
 class QiDespot1InputSpec(CommandLineInputSpec):
-    
     # Inputs
-    param_file = File(desc='Parameter .json file', position=-1, argstr='< %s', 
+    spgr_file = File(exists=True, argstr='%s', mandatory=True,
+        position=0, desc='Path to SPGR data')
+
+    param_file = File(desc='Parameter .json file', position=1, argstr='< %s', 
         xor=['param_dict'], mandatory=True, exists=True)
 
-    param_dict = traits.Dict(desc='dictionary trait', position=-1, 
+    param_dict = traits.Dict(desc='dictionary trait', position=1, 
         argstr='', mandatory=True, xor=['param_file'])
-        
-    in_file = File(exists=True, argstr='%s', mandatory=True,
-        position=-2, desc='Path to SPGR data')
-    
+
     # Options
     verbose = traits.Bool(desc='Print more information', argstr='-v')
     threads = traits.Int(desc='Use N threads (default=4, 0=hardware limit)', argstr='--threads=%d')
@@ -99,21 +98,96 @@ class QiDespot1(CommandLine):
         
         return outputs
 
+############################ qidespot1hifi ############################
+
+class QiDespot1HifiInputSpec(CommandLineInputSpec):
+    # Inputs
+    spgr_file = File(exists=True, argstr='%s', mandatory=True,
+        position=0, desc='Path to SPGR data')
+
+    mprage_file = File(exists=True, argstr='%s', mandatory=True,
+        position=0, desc='Path to MPRAGE data')
+
+    param_file = File(desc='Parameter .json file', position=2, argstr='< %s', 
+        xor=['param_dict'], mandatory=True, exists=True)
+
+    param_dict = traits.Dict(desc='dictionary trait', position=2, 
+        argstr='', mandatory=True, xor=['param_file'])
+
+    # Options
+    verbose = traits.Bool(desc='Print more information', argstr='-v')
+    threads = traits.Int(desc='Use N threads (default=4, 0=hardware limit)', argstr='--threads=%d')
+    prefix = traits.String(desc='Add a prefix to output filenames', argstr='--out=%s')
+    b1map_file = File(desc='B1 map (ratio) file', argstr='--B1=%s')
+    mask_file = File(desc='Only process voxels within the mask', argstr='--mask=%s')
+    residuals = traits.Bool(desc='Write out residuals for each data-point', argstr='--resids')
+    clamp_T1 = traits.Float(desc='Clamp T1 between 0 and value', argstr='--clampT1=%f')
+    environ = {'QUIT_EXT':'NIFTI_GZ'}
+
+class QiDespot1HifiOutputSpec(TraitedSpec):
+    t1_map = File(desc="Path to T1 map")
+    pd_map = File(desc="Path to PD map")
+    b1_map = File(desc="Path to B1 map")
+    residual_map = File(desc="Path to residual map")
+
+class QiDespot1Hifi(CommandLine):
+    """
+    Calculate T1 & B1 map with the DESPOT1-HIFI method
+
+    Example
+    -------
+    >>> from QUIT.nipype.relaxometry import QiDespot1Hifi
+    >>> params = {'SPGR': {'TR':5E-3, 'FA':[5,10]}, 
+                  'MPRAGE': { 'FA': 5, 'TR': 5E-3, 'TI': 0.45, 'TD': 0, 'eta': 1, 'ETL': 64, 'k0': 0 }}
+    >>> hifi = QiDespot1Hifi(prefix='nipype_', param_dict=params)
+    >>> hifi.inputs.spgr_file = 'SPGR.nii.gz'
[…]
