### 2018-05-18 ad85c77
BUG: MT-asymmetry formula was wrong
