### 2018-05-24 2d332c0
WIP: More relaxometry wrappers
