### 2018-05-24 d9bb191
BUG: Residuals for DESPOT were being scaled
This was resulting in 0/0=Inf errors in masked regions
