+++ Source/Relaxometry/qidespot2fm.cpp
@@ -217,7 +217,7 @@ int main(int argc, char **argv) {
     QI::WriteImage(apply->GetOutput(1), outPrefix + "T2" + QI::OutExt());
     QI::WriteImage(apply->GetOutput(2), outPrefix + "f0" + QI::OutExt());
     QI::WriteImage(apply->GetIterationsOutput(), outPrefix + "its" + QI::OutExt());
-    QI::WriteScaledImage(apply->GetResidualOutput(), apply->GetOutput(0), outPrefix + "residual" + QI::OutExt());
+    QI::WriteImage(apply->GetResidualOutput(), outPrefix + "residual" + QI::OutExt());
     if (resids) {
         QI::WriteScaledVectorImage(apply->GetAllResidualsOutput(), apply->GetOutput(0), outPrefix + "all_residuals" + QI::OutExt());
     }
