### 2018-05-24 ffd718e
BUG: Re-declared member variables causing size() errors
FA array redeclaration obscured the one in SPGRBase
