### 2018-05-24 26283a6
REF: Removed unused header include
