+++ Source/SSFP/MTFromEllipse.cpp
@@ -25,8 +25,8 @@ public:
     const Eigen::ArrayXd int_omega2;
     const Eigen::ArrayXd &TR;
     const Eigen::ArrayXd &Trf;
-    const double T2r;
-    const double T2f;
+    const double T2_b;
+    const double T2_f;
     const double f0_Hz;
     const bool debug;
 
@@ -40,35 +40,35 @@ public:
         typedef Eigen::Array<T, Eigen::Dynamic, 1> ArrayXT;
         const T &M0  = p[0];
         const T &F   = p[1];
-        const T &kf  = p[2];
-        const T &T1f = p[3];
-        const T &T1r = T1f;
+        const T &k_bf  = p[2];
+        const T &T1_f = p[3];
+        const T &T1_b = T1_f;
 
-        const ArrayXT E1f = (-TR/T1f).exp();
-        const Eigen::ArrayXd E2f = (-TR/T2f).exp();
-        const Eigen::ArrayXd E2f_echo = (-TR/(2.0*T2f)).exp();
-        const T kr = (F > 0.0) ? (kf / F) : T(0.0);
-        //const double T1r = 1.0; // Fixed for now
-        const ArrayXT E1r = (-TR/T1r).exp();
-        const ArrayXT fk = (-TR*(kf + kr)).exp();
+        const ArrayXT E1f = (-TR/T1_f).exp();
+        const Eigen::ArrayXd E2_f = (-TR/T2_f).exp();
+        const Eigen::ArrayXd E2_fe = (-TR/(2.0*T2_f)).exp();
+        const T k_fb = (F > 0.0) ? (k_bf / F) : T(0.0);
+        //const double T1_b = 1.0; // Fixed for now
+        const ArrayXT E1_b = (-TR/T1_b).exp();
+        const ArrayXT fk = (-TR*(k_bf + k_fb)).exp();
 
-        const double G_gauss = (T2r / sqrt(2.*M_PI))*exp(-pow(2.*M_PI*f0_Hz*T2r,2) / 2.0);
+        const double G_gauss = (T2_b / sqrt(2.*M_PI))*exp(-pow(2.*M_PI*f0_Hz*T2_b,2) / 2.0);
         const Eigen::ArrayXd WT = M_PI * int_omega2 * G_gauss; // # Product of W and Trf to save a division and multiplication
         const Eigen::ArrayXd fw = (-WT).exp();
-        const ArrayXT A = 1.0 + F - fw*E1r*(F+fk);
-        const ArrayXT B = 1.0 + fk*(F-fw*E1r*(F+1.0));
-        const ArrayXT C = F*(1.0-E1r)*(1.0-fk);
+        const ArrayXT A = 1.0 + F - fw*E1_b*(F+fk);
+        const ArrayXT B = 1.0 + fk*(F-fw*E1_b*(F+1.0));
+        const ArrayXT C = F*(1.0-E1_b)*(1.0-fk);
 
-        const ArrayXT denom = (A - B*E1f*cos(flip) - (E2f*E2f)*(B*E1f-A*cos(flip)));
-        const ArrayXT Gp = M0*E2f_echo*(sin(flip)*((1.0-E1f)*B+C))/denom;
-        const ArrayXT bp = (E2f*(A-B*E1f)*(1.0+cos(flip)))/denom;
+        const ArrayXT denom = (A - B*E1f*cos(flip) - (E2_f*E2_f)*(B*E1f-A*cos(flip)));
+        const ArrayXT Gp = M0*E2_fe*(sin(flip)*((1.0-E1f)*B+C))/denom;
+        const ArrayXT bp = (E2_f*(A-B*E1f)*(1.0+cos(flip)))/denom;
 
[…]
