### 2018-05-29 873833a
ENH: Added Z-Shim with SoS combination
