+++ Source/Perfusion/qi_zshim.cpp
@@ -1,0 +1,114 @@
+/*
+ *  qi_zshim.cpp
+ *
+ *  Copyright (c) 2018 Tobias Wood.
+ *
+ *  This Source Code Form is subject to the terms of the Mozilla Public
+ *  License, v. 2.0. If a copy of the MPL was not distributed with this
+ *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ */
+
+#include <iostream>
+#include <Eigen/Dense>
+
+#include "Util.h"
+#include "ImageIO.h"
+#include "Args.h"
+#include "ApplyTypes.h"
+#include "SequenceCereal.h"
+
+class ZShimSoS : public QI::ApplyVectorF::Algorithm {
+protected:
+    const int m_inputsize, m_zshims;
+    bool m_debug = false;
+    TOutput m_zero;
+public:
+    ZShimSoS(const int insize, const int zshims, bool debug) :
+        m_inputsize(insize), m_zshims(zshims), m_debug(debug)
+    {
+        m_zero = TOutput(this->outputSize());
+        m_zero.Fill(0.);
+    }
+
+    size_t numInputs() const override { return 1; }
+    size_t numConsts() const override { return 0; }
+    size_t numOutputs() const override { return 1; }
+    const std::vector<std::string> & names() const {
+        static std::vector<std::string> _names = {"zshim"};
+        return _names;
+    }
+    size_t dataSize() const override { return m_inputsize; }
+    size_t outputSize() const override { return m_inputsize / m_zshims; }
+    std::vector<float> defaultConsts() const override {
+        std::vector<float> def(0);
+        return def;
+    }
+    TOutput zero() const override { return m_zero; }
+    bool apply(const std::vector<TInput> &inputs, const std::vector<TConst> &/* Unused */,
+               const TIndex &, // Unused
+               std::vector<TOutput> &outputs, TOutput &/* Unused */,
+               TInput &/* Unused */, TIterations &/* Unused */) const override
+    {
+        if (m_debug) std::cout << "Input: " << inputs[0] << "\n" << "Input size: " << m_inputsize << " zshims: " << m_zshims << " rows: " << (m_inputsize / m_zshims) << std::endl;
+        Eigen::Map<const Eigen::MatrixXf> input(inputs[0].GetDataPointer(), m_zshims,  m_inputsize / m_zshims);
+        if (m_debug) std::cout << "Mapped input:\n" << input << std::endl;
+        // Eigen::Map<Eigen::ArrayXf> output(outputs[0].GetDataPointer(), m_inputsize / m_zshims);
+        Eigen::VectorXf output = input.colwise().norm();
+        if (m_debug) std::cout << "Eigen output: " << output.transpose() << std::endl;
+        if (m_debug) std::cout << "outputSize() " << this->outputSize() << std::endl;
+        if (m_debug) std::cout << "outputs " << outputs << " [0] " << outputs[0] << std::endl;
+        for (size_t i = 0; i < this->outputSize(); i++) {
+            outputs[0][i] = output[i];
+        }
+        if (m_debug) std::cout << "Output: " << outputs[0] << std::endl;
+        return true;
+    }
+};
+
+/*
+ * Main
+ */
+int main(int argc, char **argv) {
+    Eigen::initParallel();
+    args::ArgumentParser parser("Combines Z-Shimmed data.\nhttp://github.com/spinicist/QUIT");
+    args::Positional<std::string> input_path(parser, "ZSHIM_FILE", "Input Z-Shimmed file");
+    args::HelpFlag help(parser, "HELP", "Show this help message", {'h', "help"});
[…]
