### 2018-05-29 111242f
ENH: Added flexible echo-times
size() now returns Eigen::Index instead of size_t
