+++ Source/Sequences/SequenceCereal.cpp
@@ -23,6 +23,7 @@ void save(cereal::JSONOutputArchive &ar, std::shared_ptr<QI::SequenceBase> const
     else if QI_SAVE( SSFPGS )
     else if QI_SAVE( SSFPMT )
     else if QI_SAVE( MultiEcho )
+    else if QI_SAVE( MultiEchoFlex )
     else if QI_SAVE( CASL )
     else { QI_FAIL("Unimplemented save for sequence type: " << s->name()); }
     #undef QI_SAVE
@@ -42,6 +43,7 @@ void load(cereal::JSONInputArchive &ar, std::shared_ptr<QI::SequenceBase> &sb) {
     else if QI_LOAD( SSFPGS )
     else if QI_LOAD( SSFPMT )
     else if QI_LOAD( MultiEcho )
+    else if QI_LOAD( MultiEchoFlex )
     else if QI_LOAD( CASL )
     else { QI_FAIL("Unimplemented load for sequence type: " << seq_type); }
     #undef QI_LOAD
@@ -65,6 +67,7 @@ namespace QI {
     QI_READSEQ( SSFPEllipseSequence )
     QI_READSEQ( SSFPMTSequence )
     QI_READSEQ( MultiEchoSequence )
+    QI_READSEQ( MultiEchoFlexSequence )
     QI_READSEQ( MP2RAGESequence )
     QI_READSEQ( CASLSequence )
     QI_READSEQ( SequenceGroup )
