+++ Source/Perfusion/qi_ase_oef.cpp
@@ -13,7 +13,6 @@
 #include <Eigen/Dense>
 
 #include "itkDerivativeImageFilter.h"
-
 #include "Util.h"
 #include "ImageIO.h"
 #include "Args.h"
@@ -23,12 +22,13 @@
 
 class ASEAlgo : public QI::ApplyF::Algorithm {
 protected:
-    const QI::MultiEchoSequence m_sequence;
+    const std::shared_ptr<QI::MultiEchoBase> m_sequence;
     const int m_inputsize;
-    int m_linear_count;
     const double m_B0;
+    double m_Tc;
     const QI::VolumeF::SpacingType m_voxsize;
-
+    Eigen::ArrayXd m_TE; // Local copy excluding values below Tc
+    Eigen::Index m_TE0;
     // Constants for calculations
     const double kappa = 0.03; // Conversion factor
     const double gamma = 42.577e6; // Gyromagnetic Ratio
@@ -36,17 +36,33 @@ protected:
     const double Hb = 0.34 / kappa; // Hct = 0.34;
 
 public:
-    ASEAlgo(const QI::MultiEchoSequence& seq, const int inputsize, const double B0, const QI::VolumeF::SpacingType voxsize) :
+    ASEAlgo(const std::shared_ptr<QI::MultiEchoBase> &seq, const int inputsize, const double B0, const QI::VolumeF::SpacingType voxsize) :
         m_sequence(seq), m_inputsize(inputsize), m_B0(B0), m_voxsize(voxsize)
     {
         // 2.94 is dHb for an OEF of 26%
-        const double Tc = 1.5 / ((4./3.) * M_PI * gamma * B0 * delta_X0 * 0.75);
-        m_linear_count = (m_sequence.TE > Tc).count();
-        QI_DB( B0 );
-        QI_DB( Tc );
-        QI_DBVEC( m_sequence.TE );
-        QI_DBVEC( (m_sequence.TE > Tc) );
-        QI_DB( m_linear_count );
+        m_Tc = 1.5 / ((4./3.) * M_PI * gamma * B0 * delta_X0 * 0.75);
+        const int above_Tc_count = (m_sequence->TE.abs() > m_Tc).count();
+        m_TE = Eigen::ArrayXd(above_Tc_count);
+        Eigen::Index ind = 0;
+        m_TE0 = m_sequence->size();
+        for (Eigen::Index i = 0; i < m_sequence->size(); i++) {
+            if (m_sequence->TE[i] == 0) {
+                m_TE0 = i;
+            }
+            if (std::abs(m_sequence->TE[i]) > m_Tc) {
+                m_TE[ind] = std::abs(m_sequence->TE[i]);
+                ind++;
+            }
+        }
+        if (m_TE0 == m_sequence->size()) {
+            QI_FAIL("Did not find a zero echo-time in input");
+        }
+        // QI_DB( B0 );
+        // QI_DB( m_Tc );
+        // QI_DB( m_TE0 );
+        // QI_DB( above_Tc_count );
+        // QI_DBVEC( m_sequence->TE );
+        // QI_DBVEC( m_TE );
     }
 
     size_t numInputs() const override  { return 1; }
@@ -94,47 +110,36 @@ public:
         }
     }
     
-    bool apply(const std::vector<TInput> &inputs, const std::vector<TConst> &consts,
+    bool apply(const std::vector<TInput> &inputs, const std::vector<TConst> &/* Unused */,
                const TIndex & /* Unused */,
                std::vector<TOutput> &outputs, TOutput &residual,
                TInput &resids, TIterations &its) const override
     {
-        const Eigen::Map<const Eigen::ArrayXXf> input(inputs[0].GetDataPointer(), m_sequence.size(), inputs[0].Size() / m_sequence.size());
[…]

[…]
