### 2018-05-29 22e9287
BUG: Removed scaling on allResiduals output
