+++ Source/Relaxometry/qidespot2fm.cpp
@@ -219,7 +219,7 @@ int main(int argc, char **argv) {
     QI::WriteImage(apply->GetIterationsOutput(), outPrefix + "its" + QI::OutExt());
     QI::WriteImage(apply->GetResidualOutput(), outPrefix + "residual" + QI::OutExt());
     if (resids) {
-        QI::WriteScaledVectorImage(apply->GetAllResidualsOutput(), apply->GetOutput(0), outPrefix + "all_residuals" + QI::OutExt());
+        QI::WriteVectorImage(apply->GetAllResidualsOutput(), outPrefix + "all_residuals" + QI::OutExt());
     }
     return EXIT_SUCCESS;
 }
