### 2018-05-30 9a47c62
REF: Split field gradient out into a new program
