### 2018-05-31 f41919b
BLD: Removed Ceres/ITK package specifications that sid can't satisfy
