### 2018-05-31 spinicist/QUIT#6 @spinicist
I have made progress with creating a Debian package - I can at least build and run the tests on Debian Sid. I am awaiting advice from the NeuroDebian team.

I do not want this to hold up the review. Are there any other outstanding issues?

Thanks!
