### 2018-05-31 9cc7ac2
BLD: Removed bats submodule, added it to Travis packages
