+++ .travis.yml
@@ -13,23 +13,26 @@ matrix:
       apt:
         sources:
         - ubuntu-toolchain-r-test
+        - sourceline: 'ppa:duggan/bats'
         packages:
         - gcc-5
         - g++-5
         - ninja-build
+        - bats
+        - bc
     env:
     - MATRIX_EVAL="CC=gcc-5 && CXX=g++-5"
   - os: osx
     osx_image: xcode9
 before_install:
   - eval "${MATRIX_EVAL}"
-  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update; brew install ninja; fi
+  - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then brew update; brew install ninja bats; fi
 script:
   - mkdir install
   - "./easy_build.sh -p $PWD/install"
   - cd cmake_build && ninja install && ninja package && cd ..
   - export PATH="$PWD/install/bin:$PATH"
-  - External/bats/bin/bats Test
+  - bats Test
 deploy:
   - provider: releases
     api_key:
