### 2018-05-31 3f48868
BUG: Basename was returning the leading /
