### 2018-05-31 0cf2a85
GitHub Edit Links
Fixes #12

Signed-off-by: Lewis Cowles <lewiscowles@me.com>
