### 2018-06-14 04509be
BUG: Fixed qiaffine flipping
