### 2018-06-14 610fcea
ENH: Added QI_LOG and verbose ReadImage
