+++ Source/Core/Macro.h
@@ -42,6 +42,7 @@
     exit(EXIT_FAILURE);          \
 }
 
+#define QI_LOG( vb, msg ) if ( vb ) std::cerr << msg << std::endl;
 #define QI_DB( x ) std::cout << #x << ": " << x << std::endl;
 #define QI_DBMAT( x ) std::cout << #x << "\n" << x << std::endl;
 #define QI_DBVEC( x ) std::cout << #x << ": " << x.transpose() << std::endl;
