### 2018-06-14 cd9e62e
BUG: RF Profile was not respecting the chosen slab direction
