### 2018-06-15 #656 @spinicist
Thank you very much for your time.

I will keep pursuing the NeuroDebian packages, but there is a steep learning curve there. Hopefully will have some time when I get back from ISMRM.

@cMadan do you need anything more from me?
