### 2018-06-17 #656 @cMadan
@oesteban, no problem, thank you for the thorough review!

@spinicist, you're almost done! I just need the DOI for an archived version of the current code (i.e., deposit it on Zenodo or FigShare).
