### 2018-06-18 b2e9854
DOC: Added Zenodo badge
