+++ README.md
@@ -2,6 +2,8 @@
 
 [![Build Status](https://travis-ci.org/spinicist/QUIT.svg?branch=master)](https://travis-ci.org/spinicist/QUIT)
 
+[![DOI](https://zenodo.org/badge/37066948.svg)](https://zenodo.org/badge/latestdoi/37066948)
+
 Credit / Blame / Contact - Tobias Wood - tobias.wood@kcl.ac.uk
 
 This Source Code Form is subject to the terms of the Mozilla Public
