### 2018-07-09 #656 @spinicist
@cMadan - Hi, I'm slightly concerned the meta-data for my paper is incorrect. If you look on the "Cite As:" section, it says "Wood, (2018) ...", i.e. my initials are missing. Similarly, I downloaded the paper and dragged it into Zotero and there it comes up with the author as "C Wood, Tobias", instead of "Wood, Tobias C". Could you take a quick look please?

Also does JOSS link to ORCID?
