### 2018-07-30 #656 @cMadan
Hi @spinicist, sorry I didn't see your reply here earlier!

The footer and website do just show the last name, without any initials. I'm not sure about how Zotero parses the meta data. I'm also not sure how integrated JOSS is with ORCiD (@arfon, maybe you can comment on this when you have a chance?). As far as I can tell, things are in order in the paper.md and resulting PDF.
