### 2018-07-30 #656 @arfon
It is possible that the Crossref metadata is incorrect. That's possibly/probably what's going on.

> Also does JOSS link to ORCID?

I don't know quite what you mean - we use ORCID logins but we don't have any other integrations. I believe Crossref has the ability to auto-update\ ORCID profiles: https://www.crossref.org/blog/crossref-to-auto-update-orcid-records/
