### 2018-08-03 #656 @spinicist
Thanks both. I was particularly concerned because my Google Scholar profile had not picked up the paper, but now it has.

I don't think I added my ORCID to the paper meta-data? Is it possible to do that?
