### 2018-08-03 #656 @arfon
Your ORCID was associated with this submission when we made the DOI with Crossref: https://github.com/openjournals/joss-papers/blob/master/joss.00656/10.21105.joss.00656.crossref.xml#L37

Is that the correct ORCID?
