### 2018-08-03 #656 @spinicist
Yes, that is the correct ORCID.

Of course, if I log into ORCID itself, the paper is listed there. Sorry I didn't think to do that first and taking up your time instead.

I think everything is in order now, thanks again.
